requirements.txt
# FastAPI 및 웹 서버
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# PDF 처리
PyMuPDF>=1.24.2
pymupdf4llm==0.0.10

# 데이터베이스 및 벡터 검색
supabase==2.2.0
python-dotenv==1.0.0

# 유틸리티
pydantic==2.5.0
requests==2.31.0

# 로깅 및 디버깅
loguru==0.7.2

# 개발 의존성 (선택사항)
pytest==7.4.3
pytest-asyncio==0.21.1
httpx>=0.24.0,<0.25.0